requirements.txt
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.0.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.1
peewee==3.15.3
PyJWT==2.6.0
SQLAlchemy==2.0.10
SQLAlchemy-serializer==1.4.1
typing-extensions==4.5.0
Werkzeug==2.2.2
win32-setctime==1.1.0
WTForms==3.0.1
